rename: 도메인형 디렉토리 구조 설정 HomeController로 이름 변경
