+++ src/main/java/com/donguri/jejudorang/domain/home/controller/HomeController.java
@@ -1,4 +1,4 @@
-package com.donguri.jejudorang;
+package com.donguri.jejudorang.domain.home.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -6,9 +6,9 @@ import org.springframework.web.bind.annotation.RequestMapping;
 
 @Controller
 @RequestMapping("/")
-public class test {
+public class HomeController {
     @GetMapping
-    public String testCon() {
+    public String home() {
         return "/home/home";
     }
 }
